NikMachine/requirements.txt
python-telegram-bot==20.6
flask==3.0.3
requests